requirements.txt
# Core dependencies
gcsfs>=2023.6.0
google-cloud-storage>=2.10.0
pandas>=1.5.0
numpy>=1.21.0
python-dateutil>=2.8.2
tqdm>=4.64.0
psutil>=5.9.0
requests>=2.28.0

# Web interface 
streamlit>=1.28.0

# Performance enhancements
orjson>=3.8.0